notaria_4_core/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
satcfdi==4.0.0
spacy==3.7.4
pymupdf==1.23.22
pytesseract==0.3.10
weasyprint==61.1
google-cloud-secret-manager==2.18.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
python-multipart==0.0.9
